feat : 매칭 명세표 액티비티로 교체 #17

+++ app/src/main/java/com/example/fit_i_trainer/data/service/MatchingService.kt
@@ -12,14 +12,14 @@ interface MatchingService {
 
     //매칭거절
     @PATCH("api/matching/{matchingIdx}/reject")
-    fun matchingreject(@Path ("matchingIdx") matchingIdx : Int) : Call<BaseResponse>
+    fun matchingreject(@Path ("matchingIdx") matchingIdx : Long) : Call<BaseResponse>
     //매칭수락
     @PATCH("api/matching/{matchingIdx}/accept")
-    fun matchingaccepat(@Path("matchingIdx") matchingIdx: Int) : Call<BaseResponse>
+    fun matchingaccepat(@Path("matchingIdx") matchingIdx: Long) : Call<BaseResponse>
 
     //매칭정보조회
     @GET("api/matching/{matchingIdx}")
-    fun matchinglist(@Path("matchingIdx") matchingIdx: Int) : Call<GetMatchlistResponse>
+    fun matchinglist(@Path("matchingIdx") matchingIdx: Long) : Call<GetMatchlistResponse>
     //트레이너의 매칭목록조회
     @GET("api/matching/trainer")
     fun matchingtrainer() : Call<GettrainerResponse>
